Adds DateRangePickerInputController component
